requirements.txt
fastapi
openai
pydantic
python-dotenv
uvicorn
langchain_core
langgraph
langchain_openai